Revert "[RLlib] Trainer._evaluate -> Trainer.evaluate; Also make evaluation possible w/o evaluation worker set. (#15591)"
This reverts commit 16ddab49f576e7d8354ec3f35a0031bd1b57af97.
