+++ rllib/BUILD
@@ -1201,21 +1201,6 @@ py_test(
 #
 # Tag: evaluation
 # --------------------------------------------------------------------
-
-py_test(
-    name = "evaluation/tests/test_evaluation",
-    tags = ["evaluation"],
-    size = "medium",
-    srcs = ["evaluation/tests/test_evaluation.py"]
-)
-
-py_test(
-    name = "evaluation/tests/test_rollout_worker",
-    tags = ["evaluation"],
-    size = "medium",
-    srcs = ["evaluation/tests/test_rollout_worker.py"]
-)
-
 # mysteriously times out on travis.
 #py_test(
 #    name = "evaluation/tests/test_trajectory_view_api",
@@ -1223,6 +1208,13 @@ py_test(
 #    size = "medium",
 #    srcs = ["evaluation/tests/test_trajectory_view_api.py"]
 #)
+
+py_test(
+    name = "evaluation/tests/test_rollout_worker",
+    tags = ["evaluation"],
+    size = "medium",
+    srcs = ["evaluation/tests/test_rollout_worker.py"]
+)
 
 
 # --------------------------------------------------------------------
@@ -1408,6 +1400,13 @@ py_test(
     tags = ["tests_dir", "tests_dir_E"],
     size = "medium",
     srcs = ["tests/test_env_with_subprocess.py"]
+)
+
+py_test(
+    name = "tests/test_evaluators",
+    tags = ["tests_dir", "tests_dir_E"],
+    size = "medium",
+    srcs = ["tests/test_evaluators.py"]
 )
 
 py_test(
