+++ rllib/agents/trainer.py
@@ -214,7 +214,7 @@ COMMON_CONFIG: TrainerConfigDict = {
     # Whether to run evaluation in parallel to a Trainer.train() call
     # using threading. Default=False.
     # E.g. evaluation_interval=2 -> For every other training iteration,
-    # the Trainer.train() and Trainer.evaluate() calls run in parallel.
+    # the Trainer.train() and Trainer._evaluate() calls run in parallel.
     # Note: This is experimental. Possible pitfalls could be race conditions
     # for weight synching at the beginning of the evaluation loop.
     "evaluation_parallel_to_training": False,
@@ -239,7 +239,7 @@ COMMON_CONFIG: TrainerConfigDict = {
     "evaluation_num_workers": 0,
     # Customize the evaluation method. This must be a function of signature
     # (trainer: Trainer, eval_workers: WorkerSet) -> metrics: dict. See the
-    # Trainer.evaluate() method to see the default implementation. The
+    # Trainer._evaluate() method to see the default implementation. The
     # trainer guarantees all eval workers have the latest policy state before
     # this function is called.
     "custom_eval_function": None,
@@ -704,9 +704,6 @@ class Trainer(Trainable):
             self._init(self.config, self.env_creator)
 
             # Evaluation setup.
-            self.evaluation_workers = None
-            self.evaluation_metrics = {}
-            # Do automatic evaluation from time to time.
             if self.config.get("evaluation_interval"):
                 # Update env_config with evaluation settings:
                 extra_config = copy.deepcopy(self.config["evaluation_config"])
@@ -726,16 +723,14 @@ class Trainer(Trainable):
                 })
                 logger.debug(
                     "using evaluation_config: {}".format(extra_config))
-                # Create a separate evaluation worker set for evaluation.
-                # If evaluation_num_workers=0, use the evaluation set's local
-                # worker for evaluation, otherwise, use its remote workers
-                # (parallelized evaluation).
+
                 self.evaluation_workers = self._make_workers(
                     env_creator=self.env_creator,
                     validate_env=None,
                     policy_class=self._policy_class,
                     config=evaluation_config,
                     num_workers=self.config["evaluation_num_workers"])
+                self.evaluation_metrics = {}
 
     @override(Trainable)
     def cleanup(self):
@@ -797,15 +792,8 @@ class Trainer(Trainable):
         """Subclasses should override this for custom initialization."""
         raise NotImplementedError
 
-    # TODO: (sven) Deprecate in favor of Trainer.evaluate().
     @DeveloperAPI
     def _evaluate(self) -> dict:
-        deprecation_warning(
-            "Trainer._evaluate", "Trainer.evaluate", error=False)
-        return self.evaluate()
-
-    @PublicAPI
-    def evaluate(self) -> dict:
         """Evaluates current policy under `evaluation_config` settings.
 
         Note that this default implementation does not do anything beyond
@@ -813,11 +801,9 @@ class Trainer(Trainable):
         """
         # Call the `_before_evaluate` hook.
         self._before_evaluate()
-
-        if self.evaluation_workers is not None:
-            # Sync weights to the evaluation WorkerSet.
-            self._sync_weights_to_workers(worker_set=self.evaluation_workers)
-            self._sync_filters_if_needed(self.evaluation_workers)
+        # Sync weights to the evaluation WorkerSet.
+        self._sync_weights_to_workers(worker_set=self.evaluation_workers)
+        self._sync_filters_if_needed(self.evaluation_workers)
 
         if self.config["custom_eval_function"]:
             logger.info("Running custom eval function {}".format(
@@ -830,35 +816,9 @@ class Trainer(Trainable):
         else:
             logger.info("Evaluating current policy for {} episodes.".format(
                 self.config["evaluation_num_episodes"]))
-            metrics = None
-            # No evaluation worker set ->
-            # Do evaluation using the local worker. Expect error due to the
-            # local worker not having an env.
-            if self.evaluation_workers is None:
-                try:
-                    for _ in range(self.config["evaluation_num_episodes"]):
-                        self.workers.local_worker().sample()
-                    metrics = collect_metrics(self.workers.local_worker())
-                except ValueError as e:
-                    if "RolloutWorker has no `input_reader` object" in \
-                            e.args[0]:
-                        raise ValueError(
-                            "Cannot evaluate w/o an evaluation worker set in "
-                            "the Trainer or w/o an env on the local worker!\n"
-                            "Try one of the following:\n1) Set "
-                            "`evaluation_interval` >= 0 to force creating a "
-                            "separate evaluation worker set.\n2) Set "
-                            "`create_env_on_driver=True` to force the local "
-                            "(non-eval) worker to have an environment to "
-                            "evaluate on.")
-                    else:
-                        raise e
-
-            # Evaluation worker set only has local worker.
-            elif self.config["evaluation_num_workers"] == 0:
+            if self.config["evaluation_num_workers"] == 0:
                 for _ in range(self.config["evaluation_num_episodes"]):
                     self.evaluation_workers.local_worker().sample()
-            # Evaluation worker set has n remote workers.
             else:
                 num_rounds = int(
                     math.ceil(self.config["evaluation_num_episodes"] /
@@ -873,10 +833,9 @@ class Trainer(Trainable):
                         w.sample.remote()
                         for w in self.evaluation_workers.remote_workers()
                     ])
-            if metrics is None:
-                metrics = collect_metrics(
-                    self.evaluation_workers.local_worker(),
-                    self.evaluation_workers.remote_workers())
+
+            metrics = collect_metrics(self.evaluation_workers.local_worker(),
+                                      self.evaluation_workers.remote_workers())
         return {"evaluation": metrics}
 
     @DeveloperAPI
@@ -1407,6 +1366,3 @@ class Trainer(Trainable):
             "{} is an invalid env specification. ".format(env_object) +
             "You can specify a custom env as either a class "
             "(e.g., YourEnvCls) or a registered env id (e.g., \"your_env\").")
-
-    def __repr__(self):
-        return self._name
